backend/requirements.txt
flask==3.0.0
python-dotenv>=1.0.0
flask-cors==4.0.0
flask-sqlalchemy==3.1.1
psycopg2-binary==2.9.9
gunicorn==21.2.0
openai==1.12.0
requests==2.31.0
# Project memory (hipporag_minimal in backend/hipporag_minimal)
numpy>=1.20.0
tqdm
pandas
pyarrow
python-igraph
httpx
filelock
tenacity
pydantic>=2.0
packaging
# Token counting for Director context window (agent conversation history)
tiktoken>=0.5.0
# Test runner
pytest>=7.0.0
